refractored code. removed langchain dependency
