--- app/debug.py
@@ -1,4 +1,0 @@
-from search import get_answers_for_question
-
-result = get_answers_for_question("What is the Hive")
-print(result)
